(PC-13536) [API] feat:venues: refactor serialization
Add a base model for PRO and NATIVE serialization models.
Base model added inside new module: serialization/base.py
