+++ api/src/pcapi/core/offers/factories.py
@@ -68,6 +68,7 @@ class VenueFactory(BaseFactory):
     )
     contact = factory.RelatedFactory("pcapi.core.offerers.factories.VenueContactFactory", factory_related_name="venue")
     bookingEmail = factory.Sequence("venue{}@example.net".format)
+    bannerUrl = "http://example.com/banner"
 
 
 class VirtualVenueFactory(VenueFactory):
