+++ api/tests/routes/pro/get_venue_test.py
@@ -81,8 +81,8 @@ class Returns200Test:
             "venueTypeCode": venue.venueTypeCode.name if venue.venueTypeCode else None,
             "visualDisabilityCompliant": venue.visualDisabilityCompliant,
             "withdrawalDetails": None,
-            "bannerUrl": None,
-            "bannerMeta": None,
+            "bannerUrl": venue.bannerUrl,
+            "bannerMeta": venue.bannerMeta,
         }
 
         # when
